+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 ## 1.15.0
-### 30-May-2024
+### 07-Jun-2024
 - Added logging functions [Pull Request No. 137](https://github.com/setup-your-mac/Setup-Your-Mac/pull/137); thanks, @robjschroeder!
 - Modified Microsoft Teams Message `activitySubtitle`
 - Activated main "Setup Your Mac" dialog with each `listitem`
@@ -13,6 +13,7 @@
 - Improved exit code processing for 'Welcome' dialog
 - Added pre-flight check for AC power (thanks for the suggestion, @arnoldtaw; thanks for the code, Obi-Josh!)
 - Added Variables for Prefill Email and Computer Name (thanks, @AndrewMBarnett!)
+- Improved Remote Validation error-checking
 
 ## 1.14.2
 ### 15-Feb-2024